refactor: improve codebase organization with renames, co-location, and barrel files
- Rename ambiguous snap-utils.ts to canvas-snap-utils.ts (preview) and
  timeline-snap-utils.ts (timeline) for clarity
- Move use-editor-hotkeys.ts into features/editor/hooks/ since it is
  only used by the editor feature
- Replace inline formatTime/formatFileSize in bundle-export-dialog with
  shared formatDuration (new) and formatBytes from utils
- Add barrel index.ts files to 10 features documenting their public API
  surface for cleaner future imports
