+++ src/features/editor/hooks/use-editor-hotkeys.ts
