+++ src/features/editor/index.ts
@@ -1,0 +1,7 @@
+// Editor feature — public API
+// Main UI shell for the video editor (properties panel, media sidebar, toolbar)
+
+export { useSelectionStore } from './stores/selection-store';
+export { useEditorStore } from './stores/editor-store';
+export { useClipboardStore } from './stores/clipboard-store';
+export { useDebugStore } from './stores/debug-store';
