+++ src/features/effects/index.ts
@@ -1,0 +1,5 @@
+// Effects feature — public API
+// CSS filter, glitch, halftone, and vignette effects
+
+export { effectsToCSSFilter, getGlitchEffects, getHalftoneEffect, getVignetteEffect } from './utils/effect-to-css';
+export { getGlitchFilterString, getScanlinesStyle } from './utils/glitch-algorithms';
