+++ src/features/export/index.ts
@@ -1,0 +1,5 @@
+// Export feature — public API
+// Canvas-based video export with Web Workers
+
+export { convertTimelineToComposition } from './utils/timeline-to-composition';
+export { renderSingleFrame } from './utils/client-render-engine';
