+++ src/features/media-library/index.ts
@@ -1,0 +1,5 @@
+// Media-library feature — public API
+// Media import, storage, thumbnails, and proxy generation
+
+export { useMediaLibraryStore } from './stores/media-library-store';
+export { mediaLibraryService } from './services/media-library-service';
