+++ src/features/preview/index.ts
@@ -1,0 +1,7 @@
+// Preview feature — public API
+// Video preview with playback controls, transform gizmos, and source monitor
+
+export { usePlaybackStore } from './stores/playback-store';
+export { useGizmoStore } from './stores/gizmo-store';
+export { useSourcePlayerStore } from './stores/source-player-store';
+export { resolveMediaUrl } from './utils/media-resolver';
