+++ src/features/preview/utils/canvas-snap-utils.ts
