+++ src/features/project-bundle/index.ts
@@ -1,0 +1,6 @@
+// Project-bundle feature — public API
+// Project serialization/deserialization for .freecut bundles
+
+export { exportProjectBundle, downloadBundle } from './services/bundle-export-service';
+export { importProjectBundle } from './services/bundle-import-service';
+export type { ExportProgress, ExportResult } from './types/bundle';
