+++ src/features/projects/index.ts
@@ -1,0 +1,4 @@
+// Projects feature — public API
+// Project CRUD and metadata management
+
+export { useProjectStore } from './stores/project-store';
