+++ src/features/settings/index.ts
@@ -1,0 +1,4 @@
+// Settings feature — public API
+// Application settings with localStorage persistence
+
+export { useSettingsStore } from './stores/settings-store';
