+++ src/features/timeline/index.ts
@@ -1,0 +1,6 @@
+// Timeline feature — public API
+// Multi-track timeline with drag/trim/stretch, transitions, markers, and keyframes
+
+export { useTimelineStore } from './stores/timeline-store';
+export { useZoomStore } from './stores/zoom-store';
+export { Timeline } from './components/timeline';
