+++ src/features/timeline/utils/timeline-snap-utils.ts
