+++ src/utils/time-utils.ts
@@ -70,6 +70,16 @@ export function framesToSeconds(frames: number, fps: number): number {
 }
 
 /**
+ * Format seconds as a human-readable duration string (e.g., "2m 30s")
+ */
+export function formatDuration(seconds: number): string {
+  if (seconds < 60) return `${Math.round(seconds)}s`;
+  const minutes = Math.floor(seconds / 60);
+  const remainingSeconds = Math.round(seconds % 60);
+  return `${minutes}m ${remainingSeconds}s`;
+}
+
+/**
  * Helper function to zero-pad numbers
  * @param num - Number to pad
  * @returns Zero-padded string (e.g., 5 -> "05")
